+++ ornn-web/src/components/ui/Toast.tsx
@@ -185,21 +185,5 @@ export function ToastContainer({
   );
 }
 
-export function useToast() {
-  const addToast = useToastStore((s) => s.addToast);
-  // exactOptionalPropertyTypes (#657): conditional spread on duration
-  // so we don't pass `{ duration: undefined }` to a contract that wants
-  // `duration?: number`.
-  return {
-    success: (message: string, duration?: number) =>
-      addToast({ type: "success", message, ...(duration !== undefined ? { duration } : {}) }),
-    error: (message: string, duration?: number) =>
-      addToast({ type: "error", message, ...(duration !== undefined ? { duration } : {}) }),
-    warning: (message: string, duration?: number) =>
-      addToast({ type: "warning", message, ...(duration !== undefined ? { duration } : {}) }),
-    info: (message: string, duration?: number) =>
-      addToast({ type: "info", message, ...(duration !== undefined ? { duration } : {}) }),
-    custom: (type: ToastType["type"], message: string, duration?: number) =>
-      addToast({ type, message, ...(duration !== undefined ? { duration } : {}) }),
-  };
-}
+// `useToast` lives in the sibling `Toast.helpers.ts` so this file only
+// exports components (react-refresh boundary, #888).
